ptite update

+++ libft/ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: maroard <maroard@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/11/04 16:33:18 by maroard           #+#    #+#             */
-/*   Updated: 2025/11/24 11:06:53 by maroard          ###   ########.fr       */
+/*   Updated: 2025/11/21 12:53:18 by maroard          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *str)
 {
-	unsigned int	i;
+	size_t	i;
 
 	i = 0;
 	while (str[i] != '\0')
